[inject] remove globals from inject.js, allowing multiple uses
